backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
openai
pysqlite3
psycopg2-binary
asyncpg
openai>=1.12.0
sse-starlette>=1.6.5
jieba>=0.42.1
bcrypt>=4.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4